requirements.txt
requests
python-dotenv
transformers
torch
bert-extractive-summarizer
spacy
sentencepiece
mlflow
boto3